feat: Create "s3GetSignedURL" function that takes the key of an S3 file and returns a signed file to be sent to the client.

+++ server/controllers/education.controller.js
@@ -1,4 +1,4 @@
-const { s3Upload, s3Delete } = require('../config/aws.config')
+const { s3UploadFile, s3DeleteFile, s3GetSignedURL } = require('../config/aws.config')
 const Education = require('../models/Education')
 const StatusCodeError = require('../utils/statusCodeError')
 
@@ -14,17 +14,19 @@ const educationCreate = async (req, res) => {
   const userId = req.userId
   const logo = req.files?.logo?.[0]
   const certificate = req.files?.certificate?.[0]
-
-  if (!(provider && qualification && dateFrom && userId))
-    throw new StatusCodeError(400, 'One or more of the required fields was not filled out.')
-
+  
   try {
+    if (!userId) throw new StatusCodeError(401, 'The user is not authenticated to make this request.')
+
+    if (!(provider && qualification && dateFrom))
+      throw new StatusCodeError(400, 'One or more of the required fields was not filled out.')
+
     var logoURL = logo
-      ? await s3Upload(logo.originalname, logo.path, 'education/logo')
+      ? await s3UploadFile(logo.originalname, logo.path, 'education/logo')
       : undefined
 
     var certificateURL = certificate
-      ? await s3Upload(certificate.originalname, certificate.path, 'education/certificate')
+      ? await s3UploadFile(certificate.originalname, certificate.path, 'education/certificate')
       : undefined
 
     const bullets = bulletsStr 
@@ -44,6 +46,13 @@ const educationCreate = async (req, res) => {
     })
     
     await education.save()
+
+    if (logoURL)
+      education.logoURL = await s3GetSignedURL(logoURL)
+      
+    if (certificateURL)
+      education.certificateURL = await s3GetSignedURL(certificateURL)
+
     return res.status(200).json({
       alert: {
         type: 'success',
@@ -52,8 +61,8 @@ const educationCreate = async (req, res) => {
       experience: education
     })
   } catch (error) {
-    if (logoURL) s3Delete(logoURL)
-    if (certificateURL) s3Delete(certificateURL)
+    if (logoURL) s3DeleteFile(logoURL)
+    if (certificateURL) s3DeleteFile(certificateURL)
     console.error(error)
     return res.status(error?.errorCode || 400).json({
       alert: {
@@ -69,6 +78,10 @@ const educationDeleteOne = async (req, res) => {
   const userId = req.userId
 
   try {
+    if (!userId) throw new StatusCodeError(401, 'The user is not authenticated to make this request.')
+
+    if (!eduId) throw new StatusCodeError(404, 'An ID for the experience to be deleted was not provided.')
+
     const education = await Education.findOneAndDelete({ _id: eduId, userId })
 
     if (!education) throw new StatusCodeError(404, 'Could not find a learning experience that matches the education ID and user ID.')
@@ -76,12 +89,12 @@ const educationDeleteOne = async (req, res) => {
     const { logoURL, certificateURL, provider } = education
     
     if (logoURL) {
-      const logo = await s3Delete(logoURL)
+      const logo = await s3DeleteFile(logoURL)
       if (!logo) throw new Error(`The ${provider} logo was not removed from the cloud.`)
     }
 
     if (certificateURL) {
-      const certificate = await s3Delete(certificateURL)
+      const certificate = await s3DeleteFile(certificateURL)
       if (!certificate) throw new Error('The certificate image was not removed from the cloud.')
     }
